# speed up hole detection by 400x