--- a/package.json
+++ b/package.json
@@ -15,6 +15,7 @@
   "dependencies": {
     "debug": "^4.2.0",
     "hashlru": "^2.3.0",
+    "is-zero-buffer": "^1.0.1",
     "lodash.debounce": "^4.0.8",
     "looper": "^4.0.0",
     "ltgt": "^2.2.1",