I had a hunch that `buf.every((x) => x === 0)` is probably slow, so I did some research, and ended up finding this lovely module by mafintosh: https://github.com/mafintosh/is-zero-buffer

I ran these simple benchmarks:

```js
var isZero = (buf) => buf.every(x => x === 0)

var buf = Buffer.alloc(65536)
var then = Date.now()

for (var i = 0; i < 100000; i++) {
  isZero(buf)
}

console.log(Date.now() - then)
```

```js
var isZero = require('is-zero-buffer')

var buf = Buffer.alloc(65536)
var then = Date.now()

for (var i = 0; i < 100000; i++) {
  isZero(buf)
}

console.log(Date.now() - then)
```

And the numbers are ridiculous:

- ~`99491` milliseconds
- ~`225` milliseconds

Easy win!