@arj03 2022-03-15
Can you try running this microbenchmark with random data?

I'm seeing this in jitdb:

Before:

```
  query one huge index (first run)

    ✔ | Query 1 big index (1st run) | 1199.9ms ± 56.76ms | -91.96 kB ± 162.02 kB | 5 | 

  query one huge index (second run)

    ✔ | Query 1 big index (2nd run) | 358.25ms ± 2.43ms | 93.54 kB ± 248.12 kB | 5 | 
```

After:

```
  query one huge index (first run)

    ✔ | Query 1 big index (1st run) | 1347.25ms ± 41.9ms | -41.5 kB ± 593.94 kB | 5 | 

  query one huge index (second run)

    ✔ | Query 1 big index (2nd run) | 409.53ms ± 6.4ms | -11.49 kB ± 455.79 kB | 5 | 
```