@staltz 2022-03-15
> Can you try running this microbenchmark with random data?

Sure, I'll try that.

> Records are at most 8k

I tried this microbenchmark with buffer size 8192 and the outcome was `63657 ms` versus `133 ms`, so about the same order of magnitude for the speed up.