@staltz 2022-03-15
Wow, I think you're onto something.

```js
var crypto = require('crypto');
// var isZero = require('./')
var isZero = (buf) => buf.every((x) => x === 0);

var MAX_SIZE = 8192
var TOTAL = 125000
var buffers = []
for (var i = 0; i < TOTAL; i++) {
  const size = Math.ceil(Math.random() * MAX_SIZE)
  buffers.push(crypto.randomBytes(size))
}

var then = Date.now()
for (var i = 0; i < TOTAL; i++) {
  isZero(buffers[i])
}
console.log(Date.now() - then)
```

## `every(x => x === 0)`

Ten different runs of the script:

35
35
55
43
91
55
57
35
35
35

Average 47ms

## isZeroBuf

27
18
19
39
39
38
18
39
29
38
18

Average 30ms