@staltz 2022-03-15
I was trying to think what's the explanation for this, and it seems to be that when the record is "real" (not deleted), then `every(x => x === 0)` will bail out quickly, on the 1st or 2nd byte, most of the times. However, if the record is deleted, then `x => x === 0` will have to traverse **all** the bytes, while `buffer.compare` (what is-zero-buffer uses under the hood) has some advantage in that case.

In other words, I think the conclusion is that we should use `is-zero-buffer` only when most records are deleted, while we should use `every` when most records are real. Reasonable to assume we have the latter case?