@staltz 2022-03-15
Hmmm, I wonder if there is a compromise. Some kind of `is-zero-buffer` fork that by default uses a for loop over the bytes, but when it detects at least N bytes of the first bytes are zero, then "probably" the whole buffer is zeroes, and it would use `compare` instead.

What do you think? Right now, this kind of optimization is not too important, but we can make an issue to remind ourselves, and when we are in production where we have real world logs with a bunch of deleted records, then we can macrobenchmark this better.