@arj03 2022-03-15
Yep, a fork would be best I think. The api would be something like peekSize where you could define the number of bytes to peek for zeroes, before falling back to the check is-zero-buffer does. The module is extremely simple so should be easy :)